68k problems

+++ src/sdl2.c
@@ -15,7 +15,7 @@ SDL_Texture* videoTexture;
 FramebufferPixelFormat framebuffer[XRES_FRAMEBUFFER * YRES_FRAMEBUFFER];
 
 #ifndef DISABLE_DEPTH_BUFFER
-uint16_t zBuffer[XRES_FRAMEBUFFER * YRES_FRAMEBUFFER];
+uint8_t zBuffer[XRES_FRAMEBUFFER * YRES_FRAMEBUFFER];
 #endif
 
 #ifndef DISABLE_STENCIL_BUFFER
